requirements.txt
flask
redis
faiss-cpu
numpy
beautifulsoup4
requests
cohere